refactor: split components into separate files and create a styles file
