+++ src/pages/MyPage/components/Mypage/InfoSection/subComponents/index.ts
@@ -1,0 +1,2 @@
+export { default as AccountSection } from './AccountSection';
+export { default as ManageSection } from './ManageSection';
