+++ src/pages/MyPage/components/Mypage/ProfileCard/subComponents/ButtonSection.tsx
@@ -1,0 +1,70 @@
+import React, { Dispatch } from 'react';
+import styled from 'styled-components';
+import { ReactComponent as Upload } from '@/pages/MyPage/tempAsssets/upload.svg';
+
+//
+//
+//
+
+interface ButtonSectionProps {
+  isModifying: boolean;
+  setisModifying: Dispatch<React.SetStateAction<boolean>>;
+}
+
+//
+//
+//
+
+/**
+ * 업로드 버튼, 수정하기 버튼 영역
+ * @param isModifying 수정중인 상태
+ * @param setIsModifing 수정중인 상태 set함수
+ */
+const ButtonSection = ({ isModifying, setisModifying }: ButtonSectionProps) => {
+  return (
+    <ProfileCardButtonSection>
+      <ProfileCardButton>
+        <Upload />
+      </ProfileCardButton>
+      <ProfileCardButton
+        onClick={() => {
+          setisModifying((prev: boolean) => !prev);
+        }}
+        isModifying={isModifying}
+      >
+        {isModifying ? '수정완료' : '수정하기'}
+      </ProfileCardButton>
+    </ProfileCardButtonSection>
+  );
+};
+
+//
+//
+//
+
+const ProfileCardButtonSection = styled.div`
+  display: flex;
+  padding: 0.5rem 0rem;
+  justify-content: center;
+  align-items: center;
+  gap: 0.5rem;
+  align-self: stretch;
+`;
+
+const ProfileCardButton = styled.button<{ isModifying?: boolean }>`
+  display: flex;
+  padding: 0.625rem;
+  justify-content: center;
+  align-items: center;
+  gap: 0.625rem;
+  border: none;
+  border-radius: 0.25rem;
+  background-color: ${({ isModifying, theme }) =>
+    isModifying ? theme.colors.primary100_1 : theme.colors.common.black};
+  color: ${({ theme }) => theme.colors.common.real_white};
+  cursor: pointer;
+  font-family: Pretendard;
+  font-size: ${({ theme }) => theme.fontSize.md};
+`;
+
+export default ButtonSection;
