+++ src/pages/MyPage/components/Mypage/ProfileCard/subComponents/index.ts
@@ -1,0 +1,6 @@
+export { default as ButtonSection } from './ButtonSection';
+export { default as InfoSection } from './InfoSection';
+export { default as IntroductionSection } from './IntroductionSection';
+export { default as LinksSection } from './LinksSection';
+export { default as NameSection } from './NameSection';
+export { default as ProfileImageSection } from './ProfileImageSection';
