+++ src/pages/MyPage/components/Mypage/index.ts
@@ -1,2 +1,2 @@
-export { default as InfoSection } from './InfoSection';
-export { default as ProfileCard } from './ProfileCard';
+export { default as InfoSection } from './InfoSection/InfoSection';
+export { default as ProfileCard } from './ProfileCard/ProfileCard';
